apps/ingest-worker/requirements.txt
nats-py==2.9.0
asyncpg==0.30.0
httpx==0.28.1
pydantic-settings==2.8.1
